+++ ShoppingCart/Entity/Coupon.cs
@@ -7,11 +7,11 @@ namespace ShoppingCartDemo.Entity
     public class Coupon
     {
         public Guid Id { get; set; }
-        public decimal MinimumAmount { get; set; }
-        public decimal Discount { get; set; }
+        public double MinimumAmount { get; set; }
+        public double Discount { get; set; }
         public DiscountType DiscountType { get; set; }
 
-        public Coupon(decimal minimumAmount, decimal discount, DiscountType discountType)
+        public Coupon(double minimumAmount, double discount, DiscountType discountType)
         {
             MinimumAmount = minimumAmount;
             Discount = discount;
